Generalize Capital::api() implementation to all singletons
