+++ src/SOFe/Capital/Mod.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SOFe\Capital;
+
+use SOFe\Capital\Di\FromContext;
+use SOFe\Capital\Di\Singleton;
+use SOFe\Capital\Di\SingletonArgs;
+use SOFe\Capital\Di\SingletonTrait;
+
+final class Mod implements Singleton, FromContext {
+    use SingletonArgs, SingletonTrait;
+
+    public const API_VERSION = "0.1.0";
+}
